new commit

+++ test_db.php
@@ -1,34 +1,58 @@
 <?php
 require "db.php" ;
 
-$mode = $_GET["case_label"];
-switch ($mode) {
-	case 'insert':
-		$date = date('Y-m-d',time());
-		$sql = "insert into accounts (email, fname, lname,phone, birthday,
-		gender,password) values ('yz746@njit.edu','Bongo', 'Zhao','911','$date','Male','1234');";
-		$results = runQuery($sql);
-		header("Location: index.php");
-		break;
-	case 'update':
-		$fname = 'Bongo';
-		$sql ="update accounts set password = '4321' where fname = '$fname' ";
-		$results = runQuery($sql);
-		header("Location: index.php");
-		break;
-	case 'delete':
-		$fname = 'Bongo';
-		$sql ="delete from accounts where fname = '$fname' ";
-		$results = runQuery($sql);
-		header("Location: index.php");
-		break;
-	case 'down dir':
-		header("Location: down_dir/test_dir.php");
-		break;
-	default:
-		header("Location: errors.php");
-		break;
-}
+
+
+   class User {
+      /* Member variables */
+      var $mode = $_GET["case_label"];
+       var $fname = "harris";
+       var $date = date('Y-m-d',time());
+
+     
+      /* Member functions */
+    
+      function insert(){
+      if ($mode == 'insert'){
+        $sql = "insert into accounts (email, fname, lname,phone, birthday,gender,password) values ('smv30@njit.edu','harris', 'clark','12458899','$date','Male','1244');";
+          $results = runQuery($sql);
+           $this->results."<br/>";
+       
+      }
+     
+      
+    }
+      
+    function delete(){
+      if ($mode == 'delete'){
+        $sql ="delete from accounts where fname = '$fname'";
+      
+    $results = runQuery($sql);
+       $this->results."<br/>";
+     
+      }
+    }
+      
+      function update(){
+        if ($mode == 'update'){}
+    $sql ="update accounts set password = '4321' where fname = '$fname' ";
+    $results = runQuery($sql);
+       $this->results ." <br/>";
+        
+      }
+    }
+   }
+
+
+
+$insertion= new User;
+$updation= new User;
+$deletion= new User;
+
+$insertion->insert();
+$updation->delete();
+$deletion->update();
+
 
 
 ?>
